fix: use prioritiesDescriptor for eligibility section
