+++ app/javascript/modules/listingDetails/ListingDetailsEligibility.tsx
@@ -10,6 +10,7 @@ import {
 } from "@bloom-housing/ui-components"
 import { RailsListing } from "../listings/SharedHelpers"
 import {
+  getPriorityUnits,
   isEducator,
   isEducatorOne,
   isFcfsSalesListing,
@@ -30,7 +31,6 @@ import {
 } from "../../util/languageUtil"
 import { BeforeApplyingForSale, BeforeApplyingType } from "../../components/BeforeApplyingForSale"
 import { ListingDetailsPreferences } from "./ListingDetailsPreferences"
-import type RailsUnit from "../../api/types/rails/listings/RailsUnit"
 import ErrorBoundary, { BoundaryScope } from "../../components/ErrorBoundary"
 import { ListingDetailsHMITable } from "./ListingDetailsHMITable"
 import "./ListingDetailsEligibility.scss"
@@ -54,24 +54,7 @@ export const ListingDetailsEligibility = ({
 }: ListingDetailsEligibilityProps) => {
   const isAllSRO = listingHasOnlySROUnits(listing)
   const isSomeSRO = listingHasSROUnits(listing)
-  const priorityUnits = []
-
-  listing.Units?.forEach((unit: RailsUnit) => {
-    const priorityUnit = priorityUnits?.find((priorityUnit: ReducedUnit) => {
-      return priorityUnit.name === unit.Priority_Type
-    })
-
-    if (unit.Priority_Type && !priorityUnit) {
-      priorityUnits.push({
-        name: unit.Priority_Type,
-        numberOfUnits: 1,
-      })
-    }
-
-    if (unit.Priority_Type && priorityUnit) {
-      priorityUnit.numberOfUnits++
-    }
-  })
+  const priorityUnits = listing.prioritiesDescriptor || getPriorityUnits(listing.Units)
 
   let occupancySubtitle = ""
   if (isSale(listing)) {
